Add files via upload
This is a sample SQL Database made as a final project for a Database Management Systems course.  This database was made for a fictional company, "The City of Denver Food Bank".  This was intended to be a transactional database that would track various business processes for the company.  Please not that this database was not intended to contain every transactional procedure that would be required by the company, but would serve as more of a "starting point" or "proof of concept" for them.